requirements.txt
```
```
netmiko>=4.0.0
pyyaml>=6.0
schedule>=1.2.0
